goldserver/gold_admin/filters.py
import django_filters, pytz
from gold_admin.models import *
from gold_admin.utils import *
class TransactionFilter(django_filters.FilterSet):
    created_gte = django_filters.IsoDateTimeFilter( name="created_gte", lookup_expr='gte' )
    created_lte = django_filters.IsoDateTimeFilter( name="created_lte", lookup_expr='lte' )
	#timestamp_gte = django_filters.DateTimeFilter(name="timestamp", lookup_expr='gte')
    class Meta:
        model = Transaction
        fields =[  'created_gte', 'created_lte' ]

#