goldserver/requirements.txt
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.4.1
Automat==0.6.0
certifi==2018.4.16
chardet==3.0.4
constantly==15.1.0
Django==1.11.12
django-allauth==0.35.0
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
ez-setup==0.9
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
Markdown==2.6.11
msgpack-python==0.5.6
MySQL-python==1.2.5
oauthlib==2.0.7
pathlib==1.0.1
Pillow==5.1.0
pypiwin32==223
python-dateutil==2.7.3
python-openid==2.2.5
pytz==2018.4
pywin32==223
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
txaio==2.10.0
urllib3==1.22
zope.interface==4.5.0
